+++ unicode.c
@@ -56,38 +56,38 @@ FAST_CALL unsigned int utf8in(const uint8_t *c, uchar_t *out) { /* only for inte
 
 FAST_CALL uint8_t *utf8out(uchar_t i, uint8_t *c) {
     if (i < 0x800) {
-        *c++=0xc0 | (uint8_t)(i >> 6);
-        *c++=0x80 | (i & 0x3f);
+        *c++ = (uint8_t)(0xc0 | (i >> 6));
+        *c++ = (uint8_t)(0x80 | (i & 0x3f));
         return c;
     }
     if (i < 0x10000) {
-        *c++=0xe0 | (uint8_t)(i >> 12);
-        *c++=0x80 | ((i >> 6) & 0x3f);
-        *c++=0x80 | (i & 0x3f);
+        *c++ = (uint8_t)(0xe0 | (i >> 12));
+        *c++ = (uint8_t)(0x80 | ((i >> 6) & 0x3f));
+        *c++ = (uint8_t)(0x80 | (i & 0x3f));
         return c;
     }
     if (i < 0x200000) {
-        *c++=0xf0 | (uint8_t)(i >> 18);
-        *c++=0x80 | ((i >> 12) & 0x3f);
-        *c++=0x80 | ((i >> 6) & 0x3f);
-        *c++=0x80 | (i & 0x3f);
+        *c++ = (uint8_t)(0xf0 | (i >> 18));
+        *c++ = (uint8_t)(0x80 | ((i >> 12) & 0x3f));
+        *c++ = (uint8_t)(0x80 | ((i >> 6) & 0x3f));
+        *c++ = (uint8_t)(0x80 | (i & 0x3f));
         return c;
     }
     if (i < 0x4000000) {
-        *c++=0xf8 | (i >> 24);
-        *c++=0x80 | ((i >> 18) & 0x3f);
-        *c++=0x80 | ((i >> 12) & 0x3f);
-        *c++=0x80 | ((i >> 6) & 0x3f);
-        *c++=0x80 | (i & 0x3f);
+        *c++ = (uint8_t)(0xf8 | (i >> 24));
+        *c++ = (uint8_t)(0x80 | ((i >> 18) & 0x3f));
+        *c++ = (uint8_t)(0x80 | ((i >> 12) & 0x3f));
+        *c++ = (uint8_t)(0x80 | ((i >> 6) & 0x3f));
+        *c++ = (uint8_t)(0x80 | (i & 0x3f));
         return c;
     }
     if ((i & ~(uchar_t)0x7fffffff) != 0) return c;
-    *c++=0xfc | (i >> 30);
-    *c++=0x80 | ((i >> 24) & 0x3f);
-    *c++=0x80 | ((i >> 18) & 0x3f);
-    *c++=0x80 | ((i >> 12) & 0x3f);
-    *c++=0x80 | ((i >> 6) & 0x3f);
-    *c++=0x80 | (i & 0x3f);
+    *c++ = (uint8_t)(0xfc | (i >> 30));
+    *c++ = (uint8_t)(0x80 | ((i >> 24) & 0x3f));
+    *c++ = (uint8_t)(0x80 | ((i >> 18) & 0x3f));
+    *c++ = (uint8_t)(0x80 | ((i >> 12) & 0x3f));
+    *c++ = (uint8_t)(0x80 | ((i >> 6) & 0x3f));
+    *c++ = (uint8_t)(0x80 | (i & 0x3f));
     return c;
 }
